requirements.txt
openai>=1.30.0
coqui-tts>=0.22.0
torch>=2.1.0
torchaudio>=2.1.0
openai-whisper>=20231117
streamlit>=1.35.0
python-dotenv>=1.0.0
numpy>=1.26.0
soundfile>=0.12.1
librosa>=0.10.1
pydub>=0.25.1
transformers<4.41.0
fastapi
uvicorn
python-multipart
aiofiles
audio-recorder-streamlit>=0.0.10
duckduckgo-search>=6.0.0
streamlit-authenticator>=0.3.0


